// packages/editor-core/src/model-types-primitives.d.ts
export interface TimelineMarker {
    id: string;
    time: number;
    label: string;
    color: string;
}